+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+PySight 1.3.4 (2021-02-18)
+==========================
+- Read default event-level tags from the config file.
+
+PySight 1.3.3 (2020-06-26)
+===========================
+- Switch from RawConfigParser to ConfigParser.
+- Add configuration switch to verify the SSL certificate of the FireEye iSight API.
+
+PySight 1.3.2 (2020-06-24)
+===========================
+- Don't publish events multiple times after each update. Instead, publish all newly created events at the end of the 
+script.
+
+PySight 1.3.1 (2020-06-17)
+===========================
+- Improve determination whether an IP address is a source or destination IP address.
+
 PySight 1.3 (2020-05-08)
 ===========================
 - Improvements in Threading.